python/requirements.txt
opencv-python
opencv-contrib-python
numpy
pyserial